clip_service/requirements.txt
cors
opencv-python
flask
tqdm
datasets
torch
transformers
requests
pandas
pillow